+++ compose/ui/ui/src/desktopMain/kotlin/androidx/compose/ui/window/Geometry.desktop.kt
@@ -16,15 +16,14 @@
 
 package androidx.compose.ui.window
 
-import androidx.compose.ui.unit.Dp
+import androidx.compose.ui.unit.DpInsets
 import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.DpRect
 import androidx.compose.ui.unit.DpSize
-import androidx.compose.ui.unit.IntSize
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.isFinite
 import androidx.compose.ui.unit.isSpecified
 import java.awt.Dimension
+import java.awt.Insets
 import java.awt.Point
 import java.awt.Rectangle
 import kotlin.math.roundToInt
@@ -51,40 +50,13 @@ internal fun DpSize.roundToDimension() = Dimension(
 )
 internal fun DpSize.roundToDimensionOrNull() =
     if (isSpecified) roundToDimension() else null
-internal fun DpSize.roundToIntSize() = IntSize(
-    width = width.value.roundToInt(),
-    height = height.value.roundToInt()
+
+/**
+ * Converts AWT [Insets] to [DpInsets].
+ */
+internal fun Insets.toDpInsets() = DpInsets(
+    top = top.dp,
+    left = left.dp,
+    bottom = bottom.dp,
+    right = right.dp
 )
-
-internal val DpRect.topLeft: DpOffset get() = DpOffset(left, top)
-internal operator fun DpRect.plus(offset: DpOffset): DpRect =
-    DpRect(left + offset.x, top + offset.y, right + offset.x, bottom + offset.y)
-
-internal val Dp.isReal
-    get() = isSpecified && isFinite
-
-internal fun Dp.requireReal(propertyName: String) =
-    require(isReal) { "$propertyName must be specified and finite"}
-
-internal fun DpRect.requireReal(): DpRect {
-    left.requireReal("left")
-    top.requireReal("top")
-    right.requireReal("right")
-    bottom.requireReal("bottom")
-    return this
-}
-
-internal fun DpSize.requireReal(): DpSize {
-    require(isSpecified) { "size must be specified" }
-    width.requireReal("width")
-    height.requireReal("height")
-    return this
-}
-
-internal fun DpOffset.requireReal(): DpOffset {
-    require(isSpecified) { "offset must be specified" }
-    x.requireReal("x")
-    y.requireReal("y")
-    return this
-}
-
